requirements.txt
pyserial>=3.4
pyaudio
numpy
flask
